Working nested products query
