+++ medicines/api/src/schema.rs
@@ -1,100 +1,97 @@
 use async_graphql::{Context, EmptyMutation, EmptySubscription, FieldResult, Object, Schema};
 
-use crate::{
-    azure_context::AzureContext,
-    document::{get_documents, Documents},
-    product::{
-        get_product, get_products_index, get_substance_with_products, Product, ProductIndex,
-    },
-    substance::{get_substances_index, Substance, SubstanceIndex},
-};
+use crate::{azure_context::AzureContext, products::Products};
 use search_client::models::DocumentType;
 
 pub struct QueryRoot;
 
 #[Object(desc = "Query root")]
 impl QueryRoot {
-    async fn substance(
-        &self,
-        context: &Context<'_>,
-        name: Option<String>,
-    ) -> FieldResult<Substance> {
-        let context = context.data::<AzureContext>()?;
-        match name {
-            Some(name) => get_substance_with_products(&name, &context.client)
-                .await
-                .map_err(|e| {
-                    tracing::error!("Error fetching results from Azure search service: {:?}", e);
-                    e.into()
-                }),
-            None => Err(anyhow::anyhow!(
-                "Getting a substance without providing a substance name is not supported."
-            )
-            .into()),
-        }
-    }
+    // async fn substance(
+    //     &self,
+    //     context: &Context<'_>,
+    //     name: Option<String>,
+    // ) -> FieldResult<Substance> {
+    //     let context = context.data::<AzureContext>()?;
+    //     match name {
+    //         Some(name) => get_substance_with_products(&name, &context.client)
+    //             .await
+    //             .map_err(|e| {
+    //                 tracing::error!("Error fetching results from Azure search service: {:?}", e);
+    //                 e.into()
+    //             }),
+    //         None => Err(anyhow::anyhow!(
+    //             "Getting a substance without providing a substance name is not supported."
+    //         )
+    //         .into()),
+    //     }
+    // }
 
-    async fn product(&self, _context: &Context<'_>, name: String) -> FieldResult<Product> {
-        get_product(name).await.map_err(|e| {
-            tracing::error!("Error fetching results from Azure search service: {:?}", e);
-            e.into()
-        })
-    }
+    // async fn product(&self, _context: &Context<'_>, name: String) -> FieldResult<Product> {
+    //     get_product(name).await.map_err(|e| {
+    //         tracing::error!("Error fetching results from Azure search service: {:?}", e);
+    //         e.into()
+    //     })
+    // }
 
-    async fn substances_index(
-        &self,
-        context: &Context<'_>,
-        letter: String,
-    ) -> FieldResult<Vec<SubstanceIndex>> {
-        let context = context.data::<AzureContext>()?;
-        get_substances_index(&context.client, letter.chars().next().unwrap())
-            .await
-            .map_err(|e| {
-                tracing::error!("Error fetching results from Azure search service: {:?}", e);
-                e.into()
-            })
-    }
+    // async fn substances_index(
+    //     &self,
+    //     context: &Context<'_>,
+    //     letter: String,
+    // ) -> FieldResult<Vec<SubstanceIndex>> {
+    //     let context = context.data::<AzureContext>()?;
+    //     get_substances_index(&context.client, letter.chars().next().unwrap())
+    //         .await
+    //         .map_err(|e| {
+    //             tracing::error!("Error fetching results from Azure search service: {:?}", e);
+    //             e.into()
+    //         })
+    // }
 
-    async fn products_index(
-        &self,
-        context: &Context<'_>,
-        substance: String,
-    ) -> FieldResult<Vec<ProductIndex>> {
-        let context = context.data::<AzureContext>()?;
-        get_products_index(&context.client, &substance)
-            .await
-            .map_err(|e| {
-                tracing::error!("Error fetching results from Azure search service: {:?}", e);
-                e.into()
-            })
-    }
+    // async fn products_index(
+    //     &self,
+    //     context: &Context<'_>,
+    //     substance: String,
+    // ) -> FieldResult<Vec<ProductIndex>> {
+    //     let context = context.data::<AzureContext>()?;
+    //     get_products_index(&context.client, &substance)
+    //         .await
+    //         .map_err(|e| {
+    //             tracing::error!("Error fetching results from Azure search service: {:?}", e);
+    //             e.into()
+    //         })
+    // }
 
-    async fn documents(
-        &self,
-        context: &Context<'_>,
-        search: Option<String>,
-        first: Option<i32>,
-        skip: Option<i32>,
-        after: Option<String>,
-        document_types: Option<Vec<DocumentType>>,
-    ) -> FieldResult<Documents> {
-        let context = context.data::<AzureContext>()?;
-        let offset = get_offset_or_default(skip, after, 0);
+    // async fn documents(
+    //     &self,
+    //     context: &Context<'_>,
+    //     search: Option<String>,
+    //     first: Option<i32>,
+    //     skip: Option<i32>,
+    //     after: Option<String>,
+    //     document_types: Option<Vec<DocumentType>>,
+    // ) -> FieldResult<Documents> {
+    //     let context = context.data::<AzureContext>()?;
+    //     let offset = get_offset_or_default(skip, after, 0);
 
-        get_documents(
-            &context.client,
-            search.as_deref().unwrap_or(" "),
-            first,
-            offset,
-            document_types,
-            None,
-        )
-        .await
-        .map(Into::into)
-        .map_err(|e| {
-            tracing::error!("Error fetching results from Azure search service: {:?}", e);
-            e.into()
-        })
+    //     get_documents(
+    //         &context.client,
+    //         search.as_deref().unwrap_or(" "),
+    //         first,
+    //         offset,
+    //         document_types,
+    //         None,
+    //     )
+    //     .await
+    //     .map(Into::into)
+    //     .map_err(|e| {
+    //         tracing::error!("Error fetching results from Azure search service: {:?}", e);
+    //         e.into()
+    //     })
+    // }
+
+    async fn products(&self, context: &Context<'_>) -> FieldResult<Products> {
+        Ok(Products::new())
     }
 }
